move useffect to booklist
